+++ CHANGELOG.md
@@ -8,6 +8,7 @@ We will follow [Semantic Versioning](http://semver.org/).
 ## UNRELEASED
 ### Fixed
 - Content Security Policy issues in CMS12 by adding extra parameter within BackendYoastConfig and removing unnecessary inline code within FocusKeywordAnalysis.html
+- Removed loading of `extended-list-widget.min.css` within `PageOverviewWidget`
 - `doctrine/dbal` 2.10 support by adding `DbalService` which checks for the TYPO3 major version, if 10 then use the old `fetch` methods
 - Loading state of internal linking suggestions when the bodytext is empty (e.g. new content element)